Make NativeMethods.RemoveIcon method work
The NativeMethods.RemoveIcon method, as previously written, did not work.  Instead, the `SendMessage` method had to be fixed with help from `pinvoke.net` and then all that really needs to happen is the `WM_SETICON` message has to be sent to the window twice, one with a `WPARAM` of 1 and the other with `WPARAM` of zero, and both times with `LPARAM` of zero.  The newly-dead code in the RemoveIcon method was then excised.

For reference, see:
https://stackoverflow.com/questions/18580430/hide-the-icon-from-a-wpf-window

The only issue is that is p/invoke SendMessage method signature is incorrect; it threw a DllNotFoundException when we tested.  Instead, I went with the SendMessage signature in this commit, which tests perfectly.