Create ExtensionManager.Core.Builders.Interfaces.csproj
* The purpose of this class library is to gather the interfaces for the builder-pattern methods for the various objects that are build fluently.

For example, there is a BuildNew concrete class.  It is designed to implement one or more interface.

Each interface has just a subset of the methods defined by the BuildNew class.  Each interface, therefore, exposes a code contract with the builder so that people who want to build such-and-such a category of object can only call just the methods that actually are for building the object.

The interfaces that enforce these code contracts need to be visible to all the extensions that support the various versions of VS.

Since the interfaces are identically the same, and to cut down on the bloat of the compiled extension assembly, the thinking is to segregate these interfaces into a class library, so they can be shared on the binary level, not the code level.