Create UnitTests.Helper
Following the article at: https://stackoverflow.com/questions/32626253/cant-add-shared-project-to-visual-studio-test-project

The objective is to write unit tests for some of the code in this extension, so as to have guarantees that the extension code works as expected.

This is hard to do since the bulk of the code is in the Shared Items project, ExtensionManager.Shared.   A Shared Items project cannot be referenced in a Unit Test project.

We are creating this Helper class library.  The class library does not contain any code of its own.  It merely exists in order to reference the ExtensionManager.Shared project.  Then it will in turn be referenced by the unit tests.