Rolling back from creating any unit tests whatsoever
I have to pivot, again.

At present, it is not immediately obvious how to create unit tests for a Visual Studio Extension.  Perhaps there is a method out there, but for right now, it eludes me.

One of the primary issues, is that methods under tests depend on the extension being loaded in the IDE.  And I am not sure how to arrange that.

I am sure there is a Stack Overflow out there; but, for now, I am going to just back my changes out.